requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
google-api-core==2.10.0
google-auth==2.11.0
google-cloud==0.34.0
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.48.1
grpcio-status==1.48.1
idna==3.3
mysql-connector-python==8.0.30
numpy==1.23.2
packaging==21.3
pandas==1.4.4
proto-plus==1.22.1
protobuf==3.20.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
